+++ notifications-for-collapsed-admin-menu.php
@@ -2,9 +2,9 @@
 /**
  * Plugin Name: Notifications for Collapsed Admin Menu
  * Version:     1.4
- * Plugin URI:  http://coffee2code.com/wp-plugins/notifications-for-collapsed-admin-menu/
+ * Plugin URI:  https://coffee2code.com/wp-plugins/notifications-for-collapsed-admin-menu/
  * Author:      Scott Reilly
- * Author URI:  http://coffee2code.com/
+ * Author URI:  https://coffee2code.com/
  * Text Domain: notifications-for-collapsed-admin-menu
  * License:     GPLv2 or later
  * License URI: https://www.gnu.org/licenses/gpl-2.0.html