+++ notifications-for-collapsed-admin-menu.js
@@ -1,32 +1,43 @@
-function c2c_maybe_highlight_comments_icon() {
-	var target = jQuery('.awaiting-mod:first .pending-count');
-	if (! target.length)
-		return;
-	var parent = target.parents('#menu-comments');
-	var css_class = 'collapsed-with-pending';
-	var i = target.text();
+window.addEventListener( 'load', () => {
 
-	i > 0 ?
-		parent.addClass(css_class) :
-		parent.removeClass(css_class);
-}
-function c2c_maybe_highlight_plugins_icon() {
-	var target = jQuery('.plugin-count:first');
-	var parent = target.parents('#menu-plugins');
-	var css_class = 'collapsed-with-pending';
-	var i = target.text();
+	function c2c_maybe_highlight_comments_icon() {
+		const target = document.querySelector('.awaiting-mod .pending-count');
+		if ( target === null ) {
+			return;
+		}
+		const parent = target.closest('#menu-comments');
+		const css_class = 'collapsed-with-pending';
+		const i = target.textContent;
 
-	i > 0 ?
-		parent.addClass(css_class) :
-		parent.removeClass(css_class);
-}
-function c2c_maybe_highlight() {
-	c2c_maybe_highlight_comments_icon();
-	c2c_maybe_highlight_plugins_icon();
-}
-jQuery(document).ready(function($) {
+		i > 0 ?
+			parent.classList.add(css_class) :
+			parent.classList.remove(css_class);
+	}
+
+	function c2c_maybe_highlight_plugins_icon() {
+		const target = document.querySelector('.plugin-count');
+		const parent = target.closest('#menu-plugins');
+		const css_class = 'collapsed-with-pending';
+		const i = target.textContent;
+
+		i > 0 ?
+			parent.classList.add(css_class) :
+			parent.classList.remove(css_class);
+	}
+
+	function c2c_maybe_highlight() {
+		c2c_maybe_highlight_comments_icon();
+		c2c_maybe_highlight_plugins_icon();
+	}
+
 	c2c_maybe_highlight();
-	$('#awaiting-mod').ajaxSuccess(function() {
-		setTimeout('c2c_maybe_highlight()', 1000);
+
+	const observer = new MutationObserver( mutation => {
+		c2c_maybe_highlight();
 	});
+	const nfcam_observer_config = { childList: true };
+
+	observer.observe(document.querySelector('.awaiting-mod .pending-count'), nfcam_observer_config);
+	observer.observe(document.querySelector('.plugin-count'), nfcam_observer_config);
+
 });