+++ CHANGELOG.md
@@ -8,6 +8,7 @@
     * Change: Move `tests/` to `phpunit/tests/`
     * Change: Rename `phpunit.xml` to `phpunit.xml.dist` per best practices
 * Change: Note compatibility through WP 5.5+
+* New: Add FAQ indicating presence of unit tests
 
 ## 1.4.1 _(2020-05-26)_
 * New: Add TODO.md and move existing TODO list from top of main plugin file into it (and add to it)