+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 * Unit tests:
     * New: Add tests for filter `c2c_collapsed_admin_menu_icon_highlight_color`
     * New: Add test for default value for `get_bg_color()`
+    * Change: Test that JS script is both enqueued and registered
 
 ## 1.4.1 _(2020-05-26)_
 * New: Add TODO.md and move existing TODO list from top of main plugin file into it (and add to it)